feat: notification schema and refactoring
