--- contracts/kleros.json
@@ -1,310 +1,0 @@
-[
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "_disputeID",
-				"type": "uint256"
-			}
-		],
-		"name": "disputeStatus",
-		"outputs": [
-			{
-				"name": "status",
-				"type": "uint8"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	},
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "_disputeID",
-				"type": "uint256"
-			}
-		],
-		"name": "currentRuling",
-		"outputs": [
-			{
-				"name": "ruling",
-				"type": "uint256"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	},
-	{
-		"constant": false,
-		"inputs": [
-			{
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"name": "_extraData",
-				"type": "bytes"
-			}
-		],
-		"name": "appeal",
-		"outputs": [],
-		"payable": true,
-		"stateMutability": "payable",
-		"type": "function"
-	},
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "_disputeID",
-				"type": "uint256"
-			}
-		],
-		"name": "appealPeriod",
-		"outputs": [
-			{
-				"name": "start",
-				"type": "uint256"
-			},
-			{
-				"name": "end",
-				"type": "uint256"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	},
-	{
-		"constant": false,
-		"inputs": [
-			{
-				"name": "_choices",
-				"type": "uint256"
-			},
-			{
-				"name": "_extraData",
-				"type": "bytes"
-			}
-		],
-		"name": "createDispute",
-		"outputs": [
-			{
-				"name": "disputeID",
-				"type": "uint256"
-			}
-		],
-		"payable": true,
-		"stateMutability": "payable",
-		"type": "function"
-	},
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"name": "_extraData",
-				"type": "bytes"
-			}
-		],
-		"name": "appealCost",
-		"outputs": [
-			{
-				"name": "fee",
-				"type": "uint256"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	},
-	{
-		"constant": true,
-		"inputs": [
-			{
-				"name": "_extraData",
-				"type": "bytes"
-			}
-		],
-		"name": "arbitrationCost",
-		"outputs": [
-			{
-				"name": "fee",
-				"type": "uint256"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"indexed": true,
-				"name": "_arbitrable",
-				"type": "address"
-			}
-		],
-		"name": "DisputeCreation",
-		"type": "event"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"indexed": true,
-				"name": "_arbitrable",
-				"type": "address"
-			}
-		],
-		"name": "AppealPossible",
-		"type": "event"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"indexed": true,
-				"name": "_arbitrable",
-				"type": "address"
-			}
-		],
-		"name": "AppealDecision",
-		"type": "event"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"name": "_period",
-				"type": "uint8"
-			}
-		],
-		"name": "NewPeriod",
-		"type": "event"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_address",
-				"type": "address"
-			},
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"name": "_appeal",
-				"type": "uint256"
-			},
-			{
-				"name": "_voteID",
-				"type": "uint256"
-			}
-		],
-		"name": "Draw",
-		"type": "event"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"name": "_address",
-				"type": "address"
-			},
-			{
-				"indexed": true,
-				"name": "_disputeID",
-				"type": "uint256"
-			},
-			{
-				"name": "_tokenAmount",
-				"type": "int256"
-			},
-			{
-				"name": "_ETHAmount",
-				"type": "int256"
-			}
-		],
-		"name": "TokenAndETHShift",
-		"type": "event"
-	},
-	{
-		"constant": true,
-		"inputs": [
-		  {
-			"name": "",
-			"type": "uint256"
-		  }
-		],
-		"name": "disputes",
-		"outputs": [
-		  {
-			"name": "subcourtID",
-			"type": "uint96"
-		  },
-		  {
-			"name": "arbitrated",
-			"type": "address"
-		  },
-		  {
-			"name": "numberOfChoices",
-			"type": "uint256"
-		  },
-		  {
-			"name": "period",
-			"type": "uint8"
-		  },
-		  {
-			"name": "lastPeriodChange",
-			"type": "uint256"
-		  },
-		  {
-			"name": "drawsInRound",
-			"type": "uint256"
-		  },
-		  {
-			"name": "commitsInRound",
-			"type": "uint256"
-		  },
-		  {
-			"name": "ruled",
-			"type": "bool"
-		  }
-		],
-		"payable": false,
-		"stateMutability": "view",
-		"type": "function"
-	}
-]
